change: use DOMRect instead of ImageRef for bounds
The rectangle is tracked by a ResizeObserver and updates the bounds, when the browser window and image resizes. This way the hotspots are always correctly repositioned. The rect is also used to calculate the initial position of the hotspot.

This change also fixes the bug, that hotspot couldn't be changed after initial save